스프링4.1.5 와 hibernate 4.3.8 Final CRUD 와 count, list 테스트

+++ src/main/java/org/hyunkyo/dao/UserDao.java
@@ -14,7 +14,7 @@ public interface UserDao {
 
 	public User getByUserId(String user_id);
 
-	public int getUserCnt();
+	public Long getUserCnt();
 
 	public List<User> getUserList();
